refactor: replace cce tests to separate directory
